// src/app/core/entity/model/card/tense/grammatical-person.enum.ts
export enum GrammaticalPerson {
  SINGULAR = 'Singular',
  PLURAL = 'Plural',

  FIRST_PERSON_SINGULAR = 'First person singular',
  SECOND_PERSON_SINGULAR = 'Second person singular',
  THIRD_PERSON_SINGULAR = 'Third person singular',

  FIRST_PERSON_PLURAL = 'First person plural',
  SECOND_PERSON_PLURAL = 'Second person plural',
  THIRD_PERSON_PLURAL = 'Third person plural',
}